+++ pkg/resource/logging-agents-toggle/reconciler.go
@@ -3,6 +3,7 @@ package loggingagentstoggle
 import (
 	"context"
 	"reflect"
+	"time"
 
 	"github.com/blang/semver"
 	appv1 "github.com/giantswarm/apiextensions-application/api/v1alpha1"
@@ -27,26 +28,19 @@ type Reconciler struct {
 	Scheme *runtime.Scheme
 }
 
-func (r *Reconciler) GetObservabilityBundleVersion(ctx context.Context, lc loggedcluster.Interface) (semver.Version, error) {
-	// Get observability bundle app metadata.
-	appMeta := common.ObservabilityBundleAppMeta(lc)
-
-	// Retrieve the app.
-	var currentApp appv1.App
-	err := r.Client.Get(ctx, types.NamespacedName{Name: appMeta.GetName(), Namespace: appMeta.GetNamespace()}, &currentApp)
-	if err != nil {
-		return semver.Version{}, errors.WithStack(err)
-	}
-	return semver.Parse(currentApp.Spec.Version)
-}
-
 // ReconcileCreate ensure logging agents are enabled in the given cluster.
 func (r *Reconciler) ReconcileCreate(ctx context.Context, lc loggedcluster.Interface) (ctrl.Result, error) {
 	logger := log.FromContext(ctx)
 	logger.Info("Logging agents toggle create")
 
-	observabilityBundleVersion, err := r.GetObservabilityBundleVersion(ctx, lc)
+	observabilityBundleVersion, err := common.GetObservabilityBundleAppVersion(lc, r.Client, ctx)
 	if err != nil {
+		// Handle case where the app is not found.
+		if apimachineryerrors.IsNotFound(err) {
+			logger.Info("logging-agents-toggle - observability bundle app not found, requeueing")
+			// If the app is not found we should requeue and try again later (5 minutes is the app platform default reconciliation time)
+			return ctrl.Result{RequeueAfter: time.Duration(5 * time.Minute)}, nil
+		}
 		return ctrl.Result{}, errors.WithStack(err)
 	}
 
@@ -91,9 +85,24 @@ func (r *Reconciler) ReconcileDelete(ctx context.Context, lc loggedcluster.Inter
 	logger := log.FromContext(ctx)
 	logger.Info("Logging agents toggle delete")
 
-	observabilityBundleVersion, err := r.GetObservabilityBundleVersion(ctx, lc)
+	// Get observability bundle app metadata.
+	appMeta := common.ObservabilityBundleAppMeta(lc)
+	// Retrieve the app.
+	var currentApp appv1.App
+	err := r.Client.Get(ctx, types.NamespacedName{Name: appMeta.GetName(), Namespace: appMeta.GetNamespace()}, &currentApp)
 	if err != nil {
-		return ctrl.Result{}, errors.WithStack(client.IgnoreNotFound(err))
+		// Handle case where the app is not found.
+		if apimachineryerrors.IsNotFound(err) {
+			logger.Info("logging-agents-toggle - observability bundle app not found, skipping deletion")
+			// If the app is not found we ignore the error and return, as this means the app was already deleted.
+			return ctrl.Result{}, nil
+		}
+		return ctrl.Result{}, errors.WithStack(err)
+	}
+
+	observabilityBundleVersion, err := semver.Parse(currentApp.Spec.Version)
+	if err != nil {
+		return ctrl.Result{}, errors.WithStack(err)
 	}
 
 	// Get expected configmap.